Accept whitespace before certain table tokens
Previously, tokens such as `{|`, `|+`, `|-` and `|}`
were allowed to function as tokens for parsing tables
even when they weren't at the beginning of a line, or
following whitespace at the beginning of a line. This is
incorrect.

`|` and `!` (separating headers and cells) when singular
need to be at the start of the line or after start of
line whitespace, but not `||` and `!!` which are used within
a line. `|` is a bit more complicated because the `|` token
is also used an attribute separator in tables!

Had to change a few mistaken tables. `{| |}` results in broken
tables because the `|}` is not parsed without a newline, and
our parser will consume the rest of the section or document.
Wikimedia's parser does something weird, and in this case the
empty table above (without a proper end token, so it's parsed
wonkily) moves forward in text until it encounters the end of the
document or another table and then puts an empty table. So
if you have a string after the text above, the string is *before*
the empty HTML-table. It's weird, and we will at least for now
keep this edgecase as it is.