Generator -> simpler Iterator
typing.Generator is the more general type, and
it is used for more complex generators (using `yield` that can
take input values and return a return value.

What we usually need is an iterator, which only has a `yield`
type and doesn't take input or return anything.

Added `make ruff` because ruff was making me crazy.