Fix #205: single-bracket urls are parsed too freely
Previously, any text that looked like and
`[https://external link]` text was parsed as one, even
though there is a requirement that the first part of the
arguments inside the brackets must start without whitespace
and a URL prefix (including "https://", "http://", "ftp://"...
even "mailto:").

These urls also may not contain newlines.

This should fix it the simplest way: when encoding magic
character tokens containing these parsing entities
at the very start of parsing, we check if the external
url entity is correct and if not, use the new MAGIC_LBRACKET_CHAR
and MAGIC_RBRACKET_CHAR to replace the brackets. This is
necessary, because we loop over the text going inside-out
searching for links, external links and templates, so if the
brackets are just left in place it terminates the loop
because the text hasn't changed between iterations.

When expanding magic characters in _finalize_expand, we
do the reverse there.