Enable external url parsing in _encode only + tests
I was getting crazy trying to figure out how to mesh together
changes that would make the behavior as close to wikimedia
parsing as possible. The new tests show part of it.

The issue is that the kludge I used to make the magical encoding
or external url-like entities work was to escape the brackets with
magical characters and then return them when expanding.

However, there is still a tokenized parsing step in parser.py
that parses anything with `[ square brackets like this `, so
if you feed the 'unescaped' brackets there, it will just try to
parse them as part of an external URL anyway. We don't have
lookahead in this portion (the regex-based encoding in the previous
step does), so we can't look whether there's, say, a newline
inside the brackets, or even if there *is* a bracket to be found.

So I ended up disabling this part of parser.py. The tokenization
code is left untouched, the difference is that now "[" and "]" in
the tokenops map point to text_fn, which just inserts them into
the parse tree as strings.

And it worked! Tests pass, at least, even the new ones.