Strip newlines from end template arguments when calling
Should fix #208

This stripping is done at the point of calling because
removing the newline when in _encode, for example, removes
some data we might want to keep for when we use node_to_wikitext
and revert the parse tree back to text. Then the newlines would
be gone.

Co-authored by: xxyzz