Remove newlines at the end of Lua frame args
Finally found a spot, and didn't even need to add a
.removesuffix("\n") call. When processing frames
(which are the context objects used for args when
`#invoke` is used), when cleaning up args in general
also remove *a* newline at the end of the string,
if present.  I'm removing *a* newline just to have
the same result as .removesuffix("\n") has in other
contexts.