Chaining methods in mw.html should probably return self
Fix this bug:

```lua
		local htmlTitle = mw.html.create('span')
			:attr{ id = 'coordinates' }
			:addClass( displayinline and 'noprint' or nil )
			:wikitext( "" )
```

Our implementation would break after :addClass because it
would return nil, when it should have returned `self`,
so that :wikitext() would not try to 'index' a nil value.
Even if the mw.html method doesn't succeed, it should
return `self` for this chaining; I checked the scribunto
source, and none of the methods returned without returning
something or `self`.