Move HTML tag data to Wtp context instead of global
Because we want to allow parsing HTML-like extension tags,
like `<maplink>` data, we need to add them to the HTML tag
data we have, but changing ALLOWED_HTML_TAGS and messing
around with capitalised globals doesn't seem the way to go.

ALLOWED_HTML_TAGS will stay as the base on which everything
is built, but what should be generally used in the future
instead is Wtp.allowed_html_tags, which should also include
HTMLTagData entries for extra tags.

Because a bunch of other globals rely on ALLOWED_HTML_TAGS,
these also need to be brough back inside the context object,
generated in Wtp.__init__().

There was one place with ALLOWED_HTML_TAGS left outside of
the context, in `to_wikitext` where a parse tree is turned
back into text; here, it was used only to check if a lone
tag (`<tag>` without an `</tag>` and without children inside
the element) should be outputted with a self-closing slash
(`<tag />`) or not; in practice, this should not matter.

The code in Wikitextprocessor doesn't *handle* the HTML
element, so currently adding capability to parse a
`<maplink>` element doesn't function correctly because it
outputs the contents when it apparently shouldn't.

The handling of the extension tags should probably be at a higher
level than the parser.