Test for extension tag parsing
Extension tags are HTML-looking tags in Wikitext that are
technically handled by the wikitext parser and not HTML,
but currently for our purposes I've left them being parsed
as HTML entities because they have the same kind of source
feel.

Previously, HTML-entities were parsed and validated based
on a global table of HTML-tag parsing data, but we moved
most of that handling (besides the basic core that is
extended by everything else) into the Wtp context.

By giving a extension_tags dict argument to Wtp(), you
can add (and override, I guess) the HTML tag dict with
extension tag data, and so all the HTML parsing and
'validation' is handled without extra work for just
extensions.