Implement partial #property parserfn
Currently, #statements return some raw data without
processing happening, which is what we would want
.#property to be, although #property also changes the data.

For example, coordinates are formatted, and should not
be in the form 'Point(2.294479 48.858296)'.

We might need to implement all of these if we want to be
'correct'...

This commit just takes the code from #statements, puts it in
.#property where it fits 'better' and then has #statements
called property_fn; same output of 'raw' data, which is
hopefully good enough for most purposes.