Changes to parser regex
Partial fix to #225

We can't get 100% correspondence with how Wikimedia parses
things because our parser works on different principles,
but here's a fix for an edge case. I found a few other
edge-cases (added commented-out tests for them) that I
don't think we can fix.

The regexes for the parsing-loop in core.py where we
turn links, external links, template arguments and
templates into magical characters to be stored out of
the way while doing inside-out parsing are now outside
the function def itself in global space.

The link regex has been changed to accept (in some cases)
stray square brackets inside the link itself (but not double
brackets or paired brackets).

The link regex also can now contain curly braces, etc.
because we handle checking whether the link contains
templates etc. in the regex handling function passed
into re.sub:

The link substitution function (used to handle the regex matches
we come across that match this) has been modified to look for
a combination of all the possible regexes that were moved out
into global namespace. So if the link regex lets something
through (like a template), it will spot it at that point and then
returns the original string (which means prev == current, and
the loops terminates to handle other stuff instead and returns
later to handle the link, if applicable).