Implement rudimentary #pagesize parserfn
Rarely used, but still needed for things not to break.

Implemented with a simple Sqlite query to get length of
page body. #pagesize should return bytesize, not number of
characters (depends on encoding), so to be correct
we cast body as a blob which now returns bytesize.

The Sqlite version on our machines is old enough
(a couple of years) that it doesn't implement
octet_length(), which would do this without any casting.

XXX: refactor the "name conversion + ->  namespace_id" code
in get_page into its own function; this implementation of
`#pagesize` returns the size of the first hit `WHERE title = ?`,
so it could *TECHNICALLY* return the wrong length, but
I'm not sure if this ever really ends up mattering; this is
an edge case among edgecases, and almost always we just want
PAGESIZE to return *something*.