Implement #filepath / filepath parser function
Because we don't have access to the files (and shouldn't rely
on them) and the search-function that returns a file url
based on just the file name, this is implemented with a
dummy URL that just starts with "//unimplemented/".

The double slash is a "protocol relative" url that seems to be
the default (only?) form of url returned by #filepath, based
on testing in the sandbox.

The original function has three return value variations:
- empty string when something goes wrong
- url
- PHP array [ $url, "nowiki" => true] -> Lua table ->
  Python dict {1: url, "nowiki": True}??

The last one is a mystery, but using the `nowiki` parameter
doesn't seem to affect the return string at all. I'm not sure
what it is supposed to be do, and why this happens when
an array is returned. It's possible that the first value in
the PHP array / Lua table is outputted, in which case you get
the string, and because it was all packed in a table like this
you get a 'nowiki' effect as a side-effect, and the `'nowiki'
=> true` is just documentation??