Set `package.loaded` to empty table
Fixes Wiktextract #529

Theknightwho used `package.loaded` in a new module (probably
for optimization) instead of something else like `require`,
but on our side package in general (except for a minor kludge)
is not implemented.

Because `package.loaded` is nil on our side, indexing it caused
a toooon of Lua errors.

This adds to the aforementioned kludge; `package.loaded` is
now an empty table, which should be handled more gracefully.