Implement dummy mw.ext.data.get function
mw.ext is a general namespace for Mediawiki
Lua extensions, and mw.ext.data is for
https://github.com/wikimedia/mediawiki-extensions-JsonConfig/

The extension is used to retrieve JSON data from
.tab files in (typically?) Commons space. The function
returns a table with data containing:

`{ schema = { fields: { ... },
  data = { ... },
  license = "CC0-1.0"
}`

and some optional fields.

It also should be able to return a `{ false }` table
as a negative result, but the problem we're trying
to fix (French module using mw.ext.data.get) doesn't
handle that gracefully, because it already 'knows'
what data it should be getting.

The code we're targetting with this fix accesses both
data.schema.fields and data.data with ipairs(), so
we need to go that deep and put empty tables there to
be iterated over.

Hopefully this is enough to prevent major Lua errors.