Implement some LOCAL[TIME] parserfns
Originally by LeMoussel
Co-authored by: LeMoussel

No tests because mocking datetime.now seems impossible
for now.

A better implementation might be to give Wtp
it's own static time object that can be used instead of
actual time, so that it can be set to whatever is needed.

A simpler implementation is to just always give out the same
exact time, maybe the start of the epoch.