Fix: mistaken implementation of mw.text.decode
Our implementation of mw.text.decode was broken due to
a misreading of
https://www.mediawiki.org/wiki/Extension:Scribunto/Lua_reference_manual#mw.text.decode

Here it is said that

> If boolean decodeNamedEntities is omitted or false,
> the only named entities recognized are `&lt;` (<),
> `&gt;` (>), `&amp;` (&), `&quot;` (") and `&nbsp;`
> (the non-breaking space, U+00A0).
> Otherwise, the list of HTML5 named entities to
> recognize is loaded from PHP's get_html_translation_table function.

Tatu read this as "if decodeNamedEntities is omitted or
false, decode only these values", but it turns out that
`decodeNamedEntities` is only required to decode *name
entities* that aren't in this list; all number values
are still decoded. I have no idea why this is like this,
and I had the same misunderstanding when first looking at
the documentation, but the Scribunto source code was
clearer.

This was causing a ton of weird Lua errors in etymology templates,
because there are bits of code that rely on the presence
of a *asterisk, used in historical linguistics to denote
a form of a word that is only reconstructed and not attested.

So when these entities were encoded into HTML entities (for
some reason) and then the decode failed because our implementation
was wonky, it broke the etymology template and it took me
the hole day of hunt-and-peck print debugging to finally
find where the fault was.