Link parsing: more broken link logic
If a link is in the form `[[Something#other|]]` it
is rendered as text and is broken and we will just
ignore it.

Links that are not broken:

`[[Something|]]` without # in the first arg

`[[#Other||]]` with another pipe; everything after
the first pipe is part of the second argument
for link text. It is rendered as "|"...