Fix bug with tokenizing
At some point when changing the TOKEN_RE regex, by
adding carets to make the tokenizing of items that should
appear only at the start of a line, we introduced (probably
me) a bug.

We handle multiple apostrophe (bold and italics) tokens
in a state machine loop, because they use the same basic
character. For this, we partition the strings being
tokenized, and here the problem starts: the carets
that were introduced now thought that a partitioned
string was actually at the start of a line (after a
newline) when it wasn't, it was after an italics or
bold token.

The fix is: I like the carets in the regex, reduces false
positives, so I wanted to keep them; instead, let's
use two versions of the original TOKEN_RE regex
by putting the token regex patterns in a list and
generating a pattern out of them by filtering out
the tokens that start with `^` for one of the patterns,
then use that pattern when the partitioned string is
not `i == 0` at the start of a line.

All tests pass, except one, which was written to
account for this bug.