Overload annotations with Literal[NodeKind.TEMPLATE]
Turns out that you can actually do this with the annotation
system.

WikiNode.find_child has been previously overloaded to
account for with_index, so some reduplication was necessary
to make this work.

Now, when find_child() is given NodeKind.TEMPLATE (as a
literal, not in a variable), the annotation system knows
that the function returns TemplateNodes.