[nl] Change 'ref' HTML 'category' to '*'
Continued from PR #321

There are a few wikitext-specific "HTML-looking" tags
like `includeonly` that have "*" as their category,
allowing them to be used anywhere in the text.

Dutch Wiktionary's example template "citeer" expands this kind of HTML
tags, previous code move the "div" tag outside the "ref" tag and break
the "ref" tag, make it unclosed. And "ref" is not a common HTML tag.