Heading tokens `==` should be parsed from outside in.
Wikimedia parses heading tokens from inside out:
== Foo = is parsed as  =, "= Foo ", =, with leftover
`=` characters in-between the bookends ending up in the
the heading itself. Fixes issue #352.

Fixed by in the portion of the code where we tokenize headings.
Happily, this is already done with a regex matching a whole
line, so modifying the contents of the match is easy.