Fix "00" arg name and whitespace around template arg in el edition
Trim whitespace around frame args after expanding them

Check out wikitextprocessor PR #363

Issue was that in some Greek templates, we had a ton of
`{{#if:...}}` templates as template arguments, and they
were written with spaces in between: `| {{#if:...}} {{#if:...}} |`.

The ifs left magic characters which prevented stripping
before they were expanded into empty strings, so the
trimming has to be done later.

Co-authored-by: xxyzz <gitpull@protonmail.com>