Create utility functions to check for lists and list items
This was becoming common enough in the code that I felt
that it makes sense to have these.

You could try to make some kind of subclass multiple-
inheritance thing with something like "ListNode" and
"ListItemNode"... But that feels that it would be messy.
Do you have a HTMLListNode class?? That's just the same
thing all over again. Adding this functionality into
WikiNode doesn't feel exactly correct either, although
we could create a function that returns something more
general than the contents of the `kind` field, like,
"WikiNode.functionality" -> returning an enum or literal
that tells you that it's a list or a list item...